Fix a unsubscribe race in EventLoopWorker
